server/requirements.txt
torch
torchvision
flask
pillow
numpy
opencv-python
flask_cors